main: implement basic audio segmentation
After reading raw audio data from the recording process and doing some
basic level detection on it, we can stream it directly into an encoding
process. This way, the FLAC conversion is happening on the fly, so we
don't need extra buffering or dynamic allocation, and if the system dies
partway through, we'll still have a partially written file, without any
separate backup or checkpointing mechanisms.