main: move finalization into async executor
Hmm. In some ways, this was nice and easy. In other ways, it's not as
clean as I'd like. Specifically, I'd like for the `sox(1)` child
processes to have synchronous I/O facilities, so that the main thread
can write to them from its loop, and also async `wait` facilities, so
that the finalization can happen asynchronously. There doesn't seem to
be a way to get synchronous I/O access to a child spawned from Tokio;
that much makes sense to me. But there also doesn't seem to be a way to
get Tokio to "adopt" a child spawned from the standard library so that
it can be reaped asynchronously. In principle, I think that that should
be possible; just stick it onto the usual reaper mechanisms. (More
technically, the Tokio-internal `fn spawn_child` implementations for
both Unix and Windows start by calling `std::process::Command::spawn`,
so could be refactored to just take a `std::process::Child` directly.)
For now, using `spawn_blocking` should be fine, because we should expect
the child `sox(1)` processes to exit quickly after we close their stdin.
And having more than one of these running at once should be rare.