gcs: add basic support for multipart object uploads
This has the advantage that we can set the metadata at the same time as
writing the object, which will be nice to record the sample rate and
sample count (from which a client can derive the duration).

Needs proper choice of boundary to not interfere with the content;
that's just hacked for now.